refactor: extracted deployment-specific wagmi configs
