+++ contracts/wagmi.config.hardhat.ts
@@ -1,12 +1,14 @@
 import { defineConfig } from "@wagmi/cli";
 import { hardhat } from "@wagmi/cli/plugins";
 
+// Useful for contracts which are not deployed yet
 export default defineConfig({
   out: "viem/generated.hardhat.ts",
-  contracts: [],
   plugins: [
     hardhat({
       project: ".",
+      namePrefix: "Hardhat",
+      exclude: ["Initializable.json", "UpgradedByRewrite.json"], // These artifacts crash the wagmi cli name generator
     }),
   ],
 });
